remove grpc-gateway dependency
This was causing issues when I integrated this package into the
Sourcegraph repository. We can probably be much more aggressive in the
dependencies we can remove here given we just want to parse strings...
but hopefully this is enough for now.

Test Plan: go test ./... and will successfully compile in the
sourcegraph repository.